[Feature] - Split wishlist
